Added some accounts in App

+++ src/main/java/org/example/SavingsAccount.java
@@ -1,9 +1,9 @@
 package org.example;
 
 public class SavingsAccount extends BankAccount{
-    float interestRate;
+    double interestRate;
 
-    public SavingsAccount(double balance, double minimumBalance, float interestRate, String accountHolderName) {
+    public SavingsAccount(double balance, double minimumBalance, double interestRate, String accountHolderName) {
         super(balance, minimumBalance, accountHolderName);
         this.interestRate = interestRate;
     }
